move Go-code-generating CRD methods into sep file
The end goal here is to move all the CRD methods that generate strings
of Go code to be functions in the `pkg/generate/code` package instead of
in the `pkg/model` package, which should just be for methods and structs
that *describe* the API types and fields, not for generating Go code
*from* those models.

This patch is just a temporary step to make the above goal easier. We
seprarate all the CRD methods that generate Go code into a separate file
that we can later move in a block into the `pkg/generate/code` package.
